Reuse handlers and rework test (still not working)
