+++ tests/onegov/org/test_views_dashboard.py
@@ -1,11 +1,9 @@
-import uuid
 import transaction
 
 from onegov.org.boardlets import OrgBoardlet
 from onegov.plausible.plausible_api import PlausibleAPI
-from onegov.ticket import TicketCollection
-from onegov.user import User
-from tests.onegov.ticket.test_collection import EchoHandler
+from onegov.ticket import TicketCollection, Ticket
+from tests.onegov.org.conftest import EchoHandler
 
 
 def test_view_dashboard_no_access(client):
@@ -45,55 +43,98 @@ def test_view_dashboard_tickets(session, handlers, client):
 
     transaction.begin()
 
-    editor = User(
-        id=uuid.uuid4(),
-        username='editor',
-        password='editor',
-        role='editor',
-    )
-    collection = TicketCollection(session)
+    # editor = User(
+    #     id=uuid.uuid4(),
+    #     username='editor',
+    #     password='editor',
+    #     role='editor',
+    # )
 
-    collection.open_ticket(
+    # collection = TicketCollection(session)
+    # collection.open_ticket(
+    #     handler_id='1',
+    #     handler_code='EVN',
+    #     title='Ticket After Work Beer',
+    #     group='Event',
+    #     state='open'
+    # )
+    # collection.open_ticket(
+    #     handler_id='2',
+    #     handler_code='EVN',
+    #     title='Ticket Petting Zoo',
+    #     group='Event',
+    #     state='pending'
+    # )
+    # collection.open_ticket(
+    #     handler_id='3',
+    #     handler_code='EVN',
+    #     title='Ticket Cheese Fondue',
+    #     group='Event',
+    #     state='closed'
+    # )
+
+    # transaction.commit()
+
+    session.add(Ticket(
+        number='EVN-0001',
         handler_id='1',
         handler_code='EVN',
         title='Ticket After Work Beer',
         group='Event',
-    )
-    collection.open_ticket(
+        state='open'
+    ))
+    session.add(Ticket(
+        number='EVN-0002',
         handler_id='2',
         handler_code='EVN',
         title='Ticket Petting Zoo',
         group='Event',
-    )
-    collection.open_ticket(
+        state='pending'
+    ))
+    session.add(Ticket(
+        number='EVN-0003',
         handler_id='3',
         handler_code='EVN',
         title='Ticket Cheese Fondue',
         group='Event',
-    )
+        state='closed'
+    ))
 
     transaction.commit()
+    session.flush()
+
+    # collection = TicketCollection(session)
+    # count = collection.get_count()
+    # assert count.open == 3
+    # assert count.pending == 0
+    # assert count.closed == 0
+
+    # client.login_admin()
+    # page = client.get('/dashboard')
+    # assert page.pyquery('.boardlet').length == 3
+    # assert 'Tickets' in page
+    # assert 'Zuletzt bearbeitete Seiten' in page
+    # assert 'Zuletzt bearbeitete News' in page
+    # fact_numbers = page.pyquery('.fact-number').text()
+    # fact_numbers = fact_numbers.split(' ')
+    # assert fact_numbers == ['3', '0', '3', '0', '-', '-']
+
+    # transaction.begin()
+    # collection = TicketCollection(session)
+    # # accept one ticket
+    # collection.by_handler_id('1').accept_ticket(editor)
+    # # close another ticket
+    # collection.by_handler_id('2').accept_ticket(editor)
+    # collection.by_handler_id('2').close_ticket()
+    # # the third ticket remains open
+    # transaction.commit()
 
     collection = TicketCollection(session)
-    assert collection.query().count() == 3
-    assert collection.query().filter_by(state='open').count() == 3
-
-    client.login_admin()
-    page = client.get('/dashboard')
-    fact_numbers = page.pyquery('.fact-number').text()
-    fact_numbers = fact_numbers.split(' ')
-    assert fact_numbers == ['3', '0', '3', '0', '-', '-']
-
-    transaction.begin()
-    # accept one ticket
-    collection.by_handler_id('1').accept_ticket(editor)
-    # close another ticket
-    collection.by_handler_id('2').accept_ticket(editor)
-    collection.by_handler_id('2').close_ticket()
-    # the third ticket remains open
-    transaction.commit()
-
-    collection = TicketCollection(session)
+    count = collection.get_count()
+    print(count)
+    assert count.open == 1
+    assert count.pending == 1
+    assert count.closed == 1
     assert collection.query().count() == 3
     assert collection.query().filter_by(state='open').count() == 1
     assert collection.query().filter_by(state='pending').count() == 1
@@ -101,6 +142,13 @@ def test_view_dashboard_tickets(session, handlers, client):
 
     client.login_admin()
     page = client.get('/dashboard')
+    assert page.pyquery('.open-tickets').attr('data-count') == '1'
+    assert page.pyquery('.pending-tickets').attr('data-count') == '1'
+    assert page.pyquery('.closed-tickets').attr('data-count') == '1'
+    assert page.pyquery('.boardlet').length == 3
+    assert 'Tickets' in page
+    assert 'Zuletzt bearbeitete Seiten' in page
+    assert 'Zuletzt bearbeitete News' in page
     fact_numbers = page.pyquery('.fact-number').text()
     fact_numbers = fact_numbers.split(' ')
     assert fact_numbers == ['1', '1', '3', '1', '-', '-']
