+++ tests/onegov/ticket/test_collection.py
@@ -5,75 +5,7 @@ from onegov.ticket.collection import ArchivedTicketCollection
 from onegov.user import UserCollection
 from unittest.mock import Mock
 
-
-class EchoHandler(Handler):
-
-    @property
-    def deleted(self):
-        return False
-
-    @property
-    def email(self):
-        return self.data.get('email')
-
-    @property
-    def title(self):
-        return self.data.get('title')
-
-    @property
-    def subtitle(self):
-        return self.data.get('subtitle')
-
-    @property
-    def group(self):
-        return self.data.get('group')
-
-    def get_summary(self, request):
-        return self.data.get('summary')
-
-    def get_links(self, request):
-        return self.data.get('links')
-
-
-class LimitingHandler(Handler):
-
-    @property
-    def deleted(self):
-        return False
-
-    @property
-    def title(self):
-        return 'Foo'
-
-    @property
-    def subtitle(self):
-        return '0xdeadbeef'
-
-    @property
-    def group(self):
-        return 'Bar'
-
-    @property
-    def handler_id(self):
-        return 1
-
-    @property
-    def handler_data(self):
-        return {}
-
-    @property
-    def email(self):
-        return 'foo@bar.com'
-
-    def get_summary(self, request):
-        return 'foobar'
-
-    @classmethod
-    def handle_extra_parameters(cls, session, query, extra_parameters):
-        if 'limit' in extra_parameters:
-            return query.limit(extra_parameters['limit'])
-        else:
-            return query
+from tests.onegov.org.conftest import EchoHandler, LimitingHandler
 
 
 class ABCTicket(Ticket):
